requirements.txt
langchain
langgraph
langchain-community
ollama
python-dotenv
tavily-python
firecrawl-py
pydantic
streamlit